subject click  pr attendance show krna raha h
